1. premiere vague labels/requirements.txt
firebase-admin==6.5.0
google-auth==2.33.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.151.0
ollama==0.3.3
mistralai==1.2.2
